# DistanceToEvent performance