--- a/docs/src/reference/CHANGELOG.md
+++ b/docs/src/reference/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## `v1.2.4`
+
+- **improved** the performance of measuring distances from observed presence data
+
 ## `v1.2.3`
 
 - **changed** the compat entry of `MakieCore` from `0.8` to `0.8, 0.9`