--- a/docs/src/reference/CHANGELOG.md
+++ b/docs/src/reference/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.3.1`
+
+- **added** support for nightly in github actions
+- **improved** the performance of pseudo-absence distance to event by making it threaded
+
 ## `v1.3.0`
 
 - **changed** Julia requirement to LTS