--- a/src/pseudoabsences.jl
+++ b/src/pseudoabsences.jl
@@ -137,7 +137,8 @@ end
     pseudoabsencemask(::Type{DistanceToEvent}, presence::T; distance::Number=100.0) where {T <: SimpleSDMLayer}
 
 Generates a mask for pseudo-absences where pseudo-absences can be within a
-`distance` (in kilometers) of the original observation.
+`distance` (in kilometers) of the original observation. Internally, this uses
+`DistanceToEvent`.
 """
 function pseudoabsencemask(
     ::Type{WithinRadius},