before

~~~
julia> @benchmark pseudoabsencemask(DistanceToEvent, presencelayer)
BenchmarkTools.Trial: 8 samples with 1 evaluation per sample.
 Range (min … max):  682.600 ms … 696.958 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     686.429 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   688.518 ms ±   5.517 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █     █   █    █ █             █                           ██  
  █▁▁▁▁▁█▁▁▁█▁▁▁▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁██ ▁
  683 ms           Histogram: frequency by time          697 ms <

 Memory estimate: 59.35 MiB, allocs estimate: 187683.
~~~

after

~~~
julia> @benchmark pseudoabsencemask(DistanceToEvent, presencelayer)
BenchmarkTools.Trial: 88 samples with 1 evaluation per sample.
 Range (min … max):  49.882 ms … 107.555 ms  ┊ GC (min … max):  0.00% … 50.97%
 Time  (median):     52.136 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   57.139 ms ±   8.205 ms  ┊ GC (mean ± σ):  10.70% ± 10.34%

   █▂ ▃                              ▁                          
  ▇████▄▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▅▅▃▄▃▄▁▄█▅█▄▅▁▄▁▁▃▁▁▁▃▁▁▁▁▁▃▁▁▁▁▃ ▁
  49.9 ms         Histogram: frequency by time         71.1 ms <

 Memory estimate: 57.67 MiB, allocs estimate: 94702.
~~~